1 Refactor #2
