+++ C#/TennisKata1/TennisKata1/TennisKata1.csproj
@@ -43,6 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Game.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TennisKataTests.cs">
       <SubType>Code</SubType>
